feat: album notes/ratings, library screens motion parity, donation icon
- Extend the note/rating/thoughts feature from songs to albums (AlbumEntity noteTitle/noteRating/noteText, AlbumMenu redesign)
- Song/album note rating precision Int -> Float (half-star support)
- IrideMotion entrance parity applied to Library Artists/Playlists/Songs and CachePlaylistScreen
- Settings: donations row gets a one-shot heart bounce/tint animation on screen entry
- DB schema 45 -> 47 (album note columns + auto migrations)

+++ app/src/main/kotlin/com/metrolist/music/db/entities/SongEntity.kt
@@ -75,7 +75,7 @@ data class SongEntity(
     @ColumnInfo(name = "noteTitle", defaultValue = "NULL")
     val noteTitle: String? = null,
     @ColumnInfo(name = "noteRating", defaultValue = "NULL")
-    val noteRating: Int? = null,
+    val noteRating: Float? = null,
     @ColumnInfo(name = "noteText", defaultValue = "NULL")
     val noteText: String? = null,
 ) {
